Move interface definitions to a single file
Importing e.g. Story (an interface) and StoryItem (a component) from
the same file was confusing. Now we can import stuff from "types",
which should make code more readable.
